fix(storage): raise the documented error kinds for malformed documents
This package's own AGENTS.md states the contract: "a caller separating
'malformed package' from 'broken reader' catches `TypeError`/`ValueError`".
Two boundaries broke it, and a caller applying that rule read a corrupt
package as an internal crash.

`check_reader_compatibility` — the one public decision point — leaked
`AttributeError` from its first attribute access when handed a parsed
mapping, `None`, or a string. The record is checked before it is read now;
malformed *contents* of a real record still degrade rather than raise,
which is a separate and deliberate rule, and is pinned as a control.

Four `from_dict` sites read a required field by raw subscript, so a
truncated document raised `KeyError` — a `LookupError`, matching neither
arm. The report named one; the sweep found four. `guards.required_field`
raises `ValueError` (the document is the right kind of thing and merely
incomplete, the distinction `mapping`'s `TypeError` already carries) and
names the record as well as the field, since a nested document surfaces
through its parent and "missing `kind`" alone does not say whose.

The sweep is a test rather than a claim, following the previous round's
lesson: it walks every `from_dict` in the package and fails on any that
reaches a required field by subscript.

Two size caps fell out. `identity.py` crossed 800, so the identifiers move
to `entity_ids.py` — split along the dependency direction rather than a
line count, since the identifiers know nothing about the collection that
indexes them and the reverse arrangement would have made the two a cycle.
Everything is re-exported, so no caller's import changes.
`test_boundary_parity.py` crossed 1200, so "how a refusal is spelled"
moves to `test_malformed_documents.py`, leaving its sibling asking only
whether a bad value is refused.
