+++ docs/contribute/plans/storage-format-v2.md
@@ -223,9 +223,10 @@ bytes/objects actually loaded.
 | `abicheck/storage/availability.py` | `FactStatus`, `Confidence`, `FactAvailability`, `AvailabilityLedger` (A0.1) |
 | `abicheck/storage/availability_status.py` | `FactStatus`, `Confidence`, `COMPARABLE_STATUSES`, `GAP_STATUSES`, `ASSERTS_NO_PRODUCER`, `STATUS_ORDER`, `CONFIDENCE_ORDER`, `worse_status`, `worse_confidence` — internal vocabulary leaf, re-exported by `availability.py` only for `FactStatus`/`Confidence` |
 | `abicheck/storage/identity.py` | `EntityKind`, `ObservationKind`, `EntityId`, `OccurrenceId`, `OccurrenceSet`, `IdentityConflict`, `elf_symbol_occurrence` (A0.2/A0.3) |
+| `abicheck/storage/entity_ids.py` | `EntityId`, `EntityKind`, `ObservationKind`, `OccurrenceId`, `elf_symbol_occurrence` — internal identifier leaf, re-exported in full by `identity.py` |
 | `abicheck/storage/canonical.py` | `canonical_form`, `canonical_json`, `semantic_digest`, `strip_capture_metadata`, `CAPTURE_METADATA_KEY` (A0.4) |
 | `abicheck/storage/versioning.py` | `PACKAGE_FORMAT_VERSION`, `COMPARISON_CONTRACT_VERSION`, `UNSTATED_VERSION`, `StorageVersions`, `ProducerIdentity`, `ReaderCompatibility`, `check_reader_compatibility` (A0.5) |
-| `abicheck/storage/guards.py` | `identity_text`, `decision_key`, `key_collection`, `provenance_text`, `diagnostics_from`, `mapping`, `enum_member`, `instance_of` — internal, not re-exported by the package |
+| `abicheck/storage/guards.py` | `identity_text`, `decision_key`, `key_collection`, `required_field`, `provenance_text`, `diagnostics_from`, `mapping`, `enum_member`, `instance_of` — internal, not re-exported by the package |
 
 `guards.py` is the one row the package does not re-export: it holds the value
 guards each module used to restate at its own doors. Three copies of one rule
@@ -275,6 +276,7 @@ user-facing, and it registers:
       - abicheck/storage/availability.py
       - abicheck/storage/availability_status.py
       - abicheck/storage/identity.py
+      - abicheck/storage/entity_ids.py
       - abicheck/storage/canonical.py
       - abicheck/storage/versioning.py
 ```
