+++ tests/unit/storage/test_landed_surface.py
@@ -57,6 +57,7 @@ _MODULES = (
 _INTERNAL_MODULES = (
     "abicheck.storage.guards",
     "abicheck.storage.availability_status",
+    "abicheck.storage.entity_ids",
 )
